requirements.txt
autopep8==1.7.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
googlemaps==4.6.0
idna==3.3
pycodestyle==2.9.1
python-dotenv==0.21.0
requests==2.28.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.12
